backend/pyproject.toml
[project]
name = "geturgame"
version = "0.1.0"
description = "Project that serves as an AI-powered tool for recommending games for Steam users, based on their library."
authors = [
    {name = "Vic K.", email = "kondratskayavictoria@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django>=5.1.5,<6.0.0",
    "django-rest-framework>=0.1.0,<0.2.0",
    "django-configurations>=2.5.1,<3.0.0",
    "psycopg2>=2.9.10,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "django-allauth>=65.3.1,<66.0.0",
    "python3-openid>=3.2.0,<4.0.0",
    "matplotlib>=3.10.0,<4.0.0",
    "textblob>=0.19.0,<0.20.0",
    "kagglehub>=0.3.6,<0.4.0",
    "joblib>=1.4.2,<2.0.0",
    "scikit-learn>=1.6.1,<2.0.0",
    "pandas>=2.2.3,<3.0.0",
    "tensorflow>=2.13.0",
    "keras>=3.8.0,<4.0.0",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pre-commit = "^4.1.0"
flake8-quotes = "^3.4.0"
black = "^24.10.0"
click = "8.0.4"

[tool.black]
line-length = 119
skip-string-normalization = false
target-version = ['py310']
